feat(parking-spot-modal): implement parking spot modal with details, actions, and tabs
