+++ resources/js/components/map/modal-parking-spot/modal-parking-spot.tsx
@@ -1,0 +1,269 @@
+import { FavoriteButton } from '@/components/frontend/button/favorite';
+import { Dialog, DialogContent, DialogDescription, DialogFooter, DialogHeader, DialogTitle } from '@/components/ui/dialog';
+import { Drawer, DrawerClose, DrawerContent, DrawerDescription, DrawerFooter, DrawerHeader, DrawerTitle } from '@/components/ui/drawer';
+import { Separator } from '@/components/ui/separator';
+import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs';
+import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from '@/components/ui/tooltip';
+import { useAuthorization } from '@/hooks/use-authorization';
+import { useMediaQuery } from '@/hooks/use-media-query';
+import { Eye, FileText, Info as InfoIcon, MapPinCheckInside, Share2, X } from 'lucide-react';
+import * as React from 'react';
+import { ActionButtons, ConfirmTab, DescriptionTab, InfoTable, MainInfo } from './modal-parts';
+import { LocationDetail } from './types';
+
+type ParkingSpotModalProps = {
+    spotId: string | null;
+    open: boolean;
+    onClose: () => void;
+    latitude: number | null;
+    longitude: number | null;
+    confirmationStatusOptions: Record<string, string>;
+};
+
+export default function ParkingSpotModal({ spotId, open, onClose, latitude, longitude, confirmationStatusOptions }: ParkingSpotModalProps) {
+    const { can, user } = useAuthorization();
+    const isLoggedIn = !!user;
+    const isDesktop = useMediaQuery('(min-width: 768px)');
+    const [data, setData] = React.useState<LocationDetail | null>(null);
+    const [loading, setLoading] = React.useState(false);
+    const [error, setError] = React.useState<string | null>(null);
+    const [copiedLocationId, setCopiedLocationId] = React.useState(false);
+    const [copiedShare, setCopiedShare] = React.useState(false);
+    const [tab, setTab] = React.useState('info');
+
+    React.useEffect(() => {
+        if (open && spotId) {
+            setLoading(true);
+            setError(null);
+            fetch(`/api/parking-spots/${spotId}`)
+                .then((res) => {
+                    if (!res.ok) throw new Error('Not found');
+                    return res.json();
+                })
+                .then(setData)
+                .catch(() => setError('Could not load location details'))
+                .finally(() => setLoading(false));
+        } else {
+            setData(null);
+            setError(null);
+        }
+    }, [spotId, open]);
+
+    React.useEffect(() => {
+        if (open) {
+            setCopiedLocationId(false);
+            setCopiedShare(false);
+            setTab('info');
+        }
+    }, [open, spotId]);
+
+    function copyLocationId(e: React.MouseEvent) {
+        e.stopPropagation();
+        if (data?.id) {
+            navigator.clipboard.writeText(data.id);
+            setCopiedLocationId(true);
+            setTimeout(() => setCopiedLocationId(false), 1400);
+        }
+    }
+
+    function getShareUrl(lat: number | null, lng: number | null, zoom = 18) {
+        if (lat !== null && lng !== null) {
+            return `${window.location.origin}${window.location.pathname}#${zoom}/${lat.toFixed(5)}/${lng.toFixed(5)}`;
+        }
+        return window.location.href;
+    }
+
+    function copyUrl() {
+        const shareUrl = getShareUrl(latitude, longitude, 18);
+        navigator.clipboard.writeText(shareUrl);
+        setCopiedShare(true);
+        setTimeout(() => setCopiedShare(false), 1400);
+    }
+
+    function LoadingSkeleton() {
+        return (
+            <div className="flex flex-col items-center justify-center gap-4 py-10">
+                <div className="mb-2 h-6 w-6 animate-spin text-orange-400">{/* Loader */}</div>
+                <div className="h-8 w-2/3 animate-pulse rounded bg-zinc-200 dark:bg-zinc-800" />
+                <div className="h-4 w-1/2 animate-pulse rounded bg-zinc-200 dark:bg-zinc-800" />
+                <div className="h-20 w-full max-w-xs animate-pulse rounded bg-zinc-100 dark:bg-zinc-900" />
+            </div>
+        );
+    }
+
+    function ErrorBlock() {
+        return (
+            <div className="flex flex-col items-center justify-center gap-4 py-10 text-red-600">
+                Could not load location details.
+                <br />
+                <button
+                    className="rounded bg-zinc-100 px-4 py-2 text-sm font-semibold text-orange-600 hover:bg-zinc-200 dark:bg-zinc-900"
+                    onClick={() => window.location.reload()}
+                >
+                    Try again
+                </button>
+            </div>
+        );
+    }
+
+    function TabBlock() {
+        if (!isLoggedIn && !data?.description)
+            return (
+                <>
+                    {data && <MainInfo data={data} />}
+                    <ActionButtons latitude={latitude} longitude={longitude} />
+                    {data && (
+                        <InfoTable
+                            data={data}
+                            isLoggedIn={isLoggedIn}
+                            isDesktop={isDesktop}
+                            copiedLocationId={copiedLocationId}
+                            setCopiedLocationId={setCopiedLocationId}
+                            copyLocationId={copyLocationId}
+                        />
+                    )}
+                </>
+            );
+        return (
+            <>
+                {data && <MainInfo data={data} />}
+                <ActionButtons latitude={latitude} longitude={longitude} />
+                <Separator />
+                <Tabs value={tab} onValueChange={setTab} className="mt-4 w-full">
+                    <TabsList className="mb-2 flex w-full">
+                        <TabsTrigger value="info" className="flex flex-1 cursor-pointer items-center justify-center gap-1">
+                            <InfoIcon className="h-4 w-4" />
+                            Info
+                        </TabsTrigger>
+                        {isLoggedIn && (
+                            <TabsTrigger value="confirm" className="flex flex-1 cursor-pointer items-center justify-center gap-1">
+                                <MapPinCheckInside className="h-4 w-4" />
+                                Confirm
+                            </TabsTrigger>
+                        )}
+                        {data?.description && (
+                            <TabsTrigger value="description" className="flex flex-1 cursor-pointer items-center justify-center gap-1">
+                                <FileText className="h-4 w-4" />
+                                Description
+                            </TabsTrigger>
+                        )}
+                    </TabsList>
+                    <TabsContent value="info">
+                        {data && (
+                            <InfoTable
+                                data={data}
+                                isLoggedIn={isLoggedIn}
+                                isDesktop={isDesktop}
+                                copiedLocationId={copiedLocationId}
+                                setCopiedLocationId={setCopiedLocationId}
+                                copyLocationId={copyLocationId}
+                            />
+                        )}
+                    </TabsContent>
+                    <TabsContent value="confirm">
+                        {isLoggedIn && data && (
+                            <ConfirmTab
+                                data={data}
+                                confirmationStatusOptions={confirmationStatusOptions}
+                                onConfirmed={() => {
+                                    fetch(`/api/parking-spots/${spotId}`)
+                                        .then((res) => res.json())
+                                        .then(setData);
+                                }}
+                            />
+                        )}
+                    </TabsContent>
+                    <TabsContent value="description">{data?.description && <DescriptionTab description={data.description} />}</TabsContent>
+                </Tabs>
+            </>
+        );
+    }
+
+    const descriptionText = 'Here you will find all the details of this parking location.';
+
+    if (isDesktop) {
+        return (
+            <Dialog open={open} onOpenChange={onClose}>
+                <DialogContent showClose={false} className="max-w-xl bg-white sm:rounded-xl dark:bg-zinc-950">
+                    <DialogHeader>
+                        <div className="flex w-full items-center justify-between gap-2">
+                            <DialogTitle className="flex items-center gap-2 text-lg font-semibold">
+                                <span className="inline-block align-middle">{/* <MapPin className="h-6 w-6 text-orange-400" /> */}</span>
+                                Parking location
+                            </DialogTitle>
+                            <div className="flex items-center gap-1 sm:gap-2">
+                                {isLoggedIn && data?.id && <FavoriteButton initial={!!data.is_favorited} id={data.id} type="parking_spot" />}
+                                <TooltipProvider>
+                                    <Tooltip open={copiedShare} onOpenChange={() => setCopiedShare(false)} delayDuration={0}>
+                                        <TooltipTrigger asChild>
+                                            <button className="cursor-pointer" aria-label="Share" onClick={copyUrl}>
+                                                <Share2 className="h-5 w-5" />
+                                            </button>
+                                        </TooltipTrigger>
+                                        <TooltipContent side="top" align="center">
+                                            Copied!
+                                        </TooltipContent>
+                                    </Tooltip>
+                                </TooltipProvider>
+                                <button className="cursor-pointer" aria-label="Close" onClick={onClose}>
+                                    <X className="h-5 w-5" />
+                                </button>
+                            </div>
+                        </div>
+                        <DialogDescription className="text-muted-foreground mb-0 text-center text-sm">{descriptionText}</DialogDescription>
+                    </DialogHeader>
+                    <div className="mt-2 max-h-[70vh] overflow-y-auto">{loading ? <LoadingSkeleton /> : error ? <ErrorBlock /> : <TabBlock />}</div>
+                    <DialogFooter className="flex flex-row justify-between gap-2 pt-4">
+                        {can('parking-spot.view') && data?.id && (
+                            <a href={route('app.parking-spots.show', { id: data.id })} target="_blank" rel="noopener">
+                                <button className="flex cursor-pointer items-center gap-2 px-4 py-2 text-sm font-semibold text-orange-600 hover:bg-orange-100">
+                                    <Eye className="h-4 w-4" />
+                                    Show
+                                </button>
+                            </a>
+                        )}
+                        <button className="cursor-pointer px-4 py-2 text-sm font-semibold text-zinc-600 dark:text-zinc-300" onClick={onClose}>
+                            Close
+                        </button>
+                    </DialogFooter>
+                </DialogContent>
+            </Dialog>
+        );
+    }
+
+    // Mobile Drawer variant
+    return (
+        <Drawer open={open} onOpenChange={onClose}>
+            <DrawerContent className="mx-auto max-w-xl bg-white dark:bg-zinc-950">
+                <DrawerHeader>
+                    <div className="flex w-full items-center justify-between gap-2">
+                        <DrawerTitle className="flex items-center gap-2 text-lg font-semibold">Parking location</DrawerTitle>
+                        <div className="flex items-center gap-1 sm:gap-2">
+                            {isLoggedIn && data?.id && <FavoriteButton initial={!!data.is_favorited} id={data.id} type="parking_spot" />}
+                            <button aria-label="Share" className="cursor-pointer" onClick={copyUrl}>
+                                <Share2 className="h-5 w-5" />
+                            </button>
+                        </div>
+                    </div>
+                </DrawerHeader>
+                <div className="max-h-[70vh] overflow-y-auto">
+                    <DrawerDescription className="text-muted-foreground mb-2 text-center text-sm">{descriptionText}</DrawerDescription>
+                    <div className="px-4 sm:px-6">{loading ? <LoadingSkeleton /> : error ? <ErrorBlock /> : <TabBlock />}</div>
+                </div>
+                <DrawerFooter className="flex flex-row gap-2">
+                    {can('parking-spot.view') && data?.id && (
+                        <a href={route('app.parking-spots.show', { id: data.id })} target="_blank" rel="noopener" className="flex-1">
+                            <button className="flex w-full items-center gap-2 px-4 py-2 text-sm font-semibold text-orange-600 hover:bg-orange-100">
+                                <Eye className="h-4 w-4" />
+                                Show
+                            </button>
+                        </a>
+                    )}
+                    <DrawerClose asChild>
+                        <button className="flex-1 px-4 py-2 text-sm font-semibold text-zinc-600 dark:text-zinc-300">Close</button>
+                    </DrawerClose>
+                </DrawerFooter>
+            </DrawerContent>
+        </Drawer>
+    );
+}
